+++ AmbermoonAndroid/Configuration.cs
@@ -87,7 +87,7 @@ namespace AmbermoonAndroid
         public GraphicFilter GraphicFilter { get; set; } = GraphicFilter.None;
         public GraphicFilterOverlay GraphicFilterOverlay { get; set; } = GraphicFilterOverlay.None;
         public Effects Effects { get; set; } = Effects.None;
-        public bool ShowPlayerStatsTooltips { get; set; } = false;
+        public bool ShowPlayerStatsTooltips { get; set; } = true;
         public bool ShowPyrdacorLogo { get; set; } = true;
         public bool ShowAdvancedLogo { get; set; } = true;
         [Obsolete("Now the fantasy intro is shown instead.")]
@@ -181,7 +181,7 @@ namespace AmbermoonAndroid
 				Names = new string[Game.NumAdditionalSavegameSlots]
 			}).ToArray();
 
-			var savegameSlots = AdditionalSavegameSlots.FirstOrDefault(s => s.GameVersionName.ToLower() == gameVersionName);
+			var savegameSlots = AdditionalSavegameSlots.FirstOrDefault(s => s.GameVersionName?.ToLower() == gameVersionName);
 
 			if (savegameSlots == null)
 			{